store instances in an arena separate from the instance IDs
for the next feature i'm planning on writing (loopers), i need the audio backend to be able to start sounds on its own. it can't create InstanceIds, because those are generated by the frontend. so i need to store instances in a way that doesn't rely on InstanceIds, and i need to optionally allow instances to be associated with an InstanceId.

i'm using generational_arena for the main instance container, since it seems like a good way to have a collection you can get indices into, and i'm using bimap to associate those indices with InstanceIds