don't loop
let's say a sequence has 2 onInterval(1.0)s in a row. the user would probably expect the sequence to wait for a total of 2 beats. but if i'm looping through the commands and completing as many of them as possible each frame, then both of those commands will be completed at the same time, thus only waiting for 1 beat. so i think in general, you'd only want to finish waiting on one command per sample.