rename sequence "commands" to "tasks"
i feel like this better encapsulates the fact that a sequence task doesn't necessarily "do" something - it can also just wait a certain amount of time. waiting is more of a "task" than a "command", which sounds more imperative.

plus, that's how i'm explaining it in the documentation, so may as well be consistent in the code