make SequenceOutputCommands separate from regular Commands
in order to implement sequence volumes, there needs to be some link between instances and sequences. the way i'm planning on making this link is adding a sequence_id field to instances. but in order to do this:
- Instance::new needs to take an Option<SequenceId>
- InstanceCommand::PlaySound needs to have an Option<SequenceId>
- anything that produces InstanceCommand::PlaySound needs to be able to include a SequenceId when needed

right now Sequences produce PlaySound commands, but they don't know their own ID. so fundamentally, Sequences should produce something else that isn't an InstanceCommand. so i've refactored SequenceOutputCommands to not include InstanceCommands