remove id fields for structs that don't need them
when i added support for setting the ID of a sound or arrangement in its settings, i needed to store the ID in the sound/arrangement so that all of the settings will follow it to the audio manager.

when i started adding ID settings to other structs, i mostly followed that pattern for consistency. however, most structs do not need to know their own ID, and it's probably better if I don't include unnecessary struct fields