remove "with playback rate change" benchmark
this benchmark was meant to test the overhead of sending a lot of commands to the backend. from what i can tell, it's not doing a very good job of this, since the time spent processing the audio overshadows the time spent processing commands, as evidenced by the fact that both benchmarks take a similar amount of time to run.