remove code related to adjusting delay times
an audio-thread-operated sample counter is not actually a good source of time for the main thread. since samples are processed at the beginning of each buffer as fast as possible, the sample count reported by the audio thread does not line up with real time. in practice, trying to adjust the delay time of tweens and instance starts just makes everything happen 10ms sooner across the board, which is pretty pointless. i'll keep investigating ways to accomplish precise timing